fix all cpp pedantic messages

+++ src/lines-as-network.cpp
@@ -95,7 +95,7 @@ Rcpp::List rcpp_lines_as_network (const Rcpp::List &sf_lines,
     {
         if (ow.size () == owb.size ())
         {
-            for (size_t i = 0; i != ow.size (); ++ i)
+            for (unsigned i = 0; i != ow.size (); ++ i)
                 if (ow [i] == "NA" && owb [i] != "NA")
                     ow [i] = owb [i];
         } else if (owb.size () > ow.size ())
